+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Modules [CV-MAP](./docs/CVMap.md), [MACRO](./docs/Macro.md), [µMAP](./docs/CVMapMicro.md), [MIDI-CAT](./docs/MidiCat.md), [MIRROR](./docs/Mirror.md), [ReMOVE Lite](./docs/ReMove.md), [X4](./docs/X4.md)
     - Added context menu option to report parameter updates to plugin-host (only in plugin-version of Rack)
 - Module [MIDI-CAT MEM](./docs/MidiCat.md#mem-expander)
+    - Added module restriction list
     - Fixed broken MIDI feedback when loading stored mappings
 - Module [TRANSIT](./docs/Transit.md)
     - Improved robustness for expander +T (e.g. crashes when using module-presets) (#412)